pyproject.toml
[tool.poetry]
name = "django-favlinks"
version = "0.1.0"
description = "A Django app to manage your favorite links or bookmarks."
authors = ["mwallraf <mwallraf@2nms.com>"]
readme = "README.md"
homepage = "https://github.com/mwallraf/django-favlinks"
license = "MIT"
packages = [{include = "django_favlinks"}]


[tool.poetry.dependencies]
python = "^3.12"
mysqlclient = "^2.2.4"
django = "^5.0.6"
pillow = "^10.3.0"
django-taggit = "^5.0.1"
django-rest-framework = "^0.1.0"
django-crispy-forms = "^2.2"
crispy-bootstrap5 = "^2024.2"

[tool.poetry.dev-dependencies]
djlint = "^1.34.1"
pytest = "^6.2"
pytest-django = "^4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

